Fix overflow risk by changing AddMarginInSats parameter from int to long
- Update interface: ILnMarketsApiService.AddMarginInSats parameter from int to long
- Update implementation: LnMarketsApiService.AddMarginInSats parameter from int to long
- Update TradeManager.cs: Remove unsafe int casting, use proper long calculations
- Update all test mocks and constants to use long instead of int
- Add explicit casting for decimal calculations to prevent overflow
- Ensure safe conversion of calculated values to long type

This prevents potential overflow when dealing with large satoshi amounts (int max ~2.1 billion vs long max ~9.2 quintillion)

+++ src/Backend/Services/LnMarketsApiService.cs
@@ -19,7 +19,7 @@ public class LnMarketsApiService : ILnMarketsApiService
         _logger = logger;
     }
 
-    public async Task<bool> AddMarginInSats(string key, string passphrase, string secret, string id, int amountInSats)
+    public async Task<bool> AddMarginInSats(string key, string passphrase, string secret, string id, long amountInSats)
     {
         var method = "POST";
         var path = "/v2/futures/add-margin";
